pyproject.toml
[project]
name = "ai-commons"
version = "0.1.0"
description = "Shared LLM interaction library"
authors = [{name="Your Name", email="your@email.com"}]
license = {text = "MIT"}
dependencies = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "typing-extensions",
]
requires-python = "^3.9"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
ai_commons = ["data/model-aliases.yaml"]

[tool.pytest.ini_options]
pythonpath = "src"